CustomSideDrawer performed.
